Feat -> Workspaces: add crud

+++ Home.tsx
@@ -1,80 +1,90 @@
-import React, { useCallback, useEffect, useState } from 'react';
+import React, { useEffect, useState } from 'react';
 import { View, StyleSheet, Dimensions, ScrollView, RefreshControl, Alert, Vibration } from 'react-native';
 import { Modal, Portal, Text, PaperProvider, TextInput, Searchbar } from 'react-native-paper';
 import { AddButton, Board } from './components';
 import { useNavigation } from '@react-navigation/native';
 import { trelloAPI } from './App';
+import { Picker } from '@react-native-picker/picker';
 
-interface HomeProps {
-  boards: {
-    id: string;
-    name: string;
-  }[],
-  setBoards: (boards: any) => void;
-}
+const containerStyle = {
+  backgroundColor: 'white',
+  padding: 10,
+  margin: 20,
+  borderRadius: 10,
+  width: Dimensions.get('window').width - 40,
+};
 
-const Home = ({
-  boards,
-  setBoards,
-}: HomeProps) => {
+const Home = () => {
   const navigation = useNavigation() as any;
 
-  const [boardss, setBoardss] = useState(boards);
-  const [visible, setVisible] = useState<boolean>(false);
+  const [boards, setBoards] = useState<any[]>([]);
+  const [workspaces, setWorkspaces] = useState<any[]>([]);
+  const [selectedWorkspace, setSelectedWorkspace] = useState<string | null>(null);
+  const [visible, setVisible] = useState<"Board" | "Workspace" | null>(null);
   const [searchQuery, setSearchQuery] = useState<string>('');
   const [refreshing, setRefreshing] = useState<boolean>(false);
+  const [edit, setEdit] = useState<string | null>(null);
 
-  const containerStyle = {
-    backgroundColor: 'white',
-    padding: 10,
-    margin: 20,
-    borderRadius: 10,
-    width: Dimensions.get('window').width - 40,
-  };
-
-  const handleAddBoard = async (name: string) =>  {
-    await trelloAPI.createBoard(name).then(() => {
-      Vibration.vibrate(200);
-    });
-    setVisible(false)
-    onRefresh();
-  }
-
-  const handleDeleteBoard = async (id: string) => {
-    await trelloAPI.deleteBoard(id).then(() => {
-      Vibration.vibrate(120);
-    });
-    onRefresh();
-  }
-
-  useEffect(() => {
-    if (searchQuery) {
-      const filteredBoards = boards.filter(board => board.name.toLowerCase().includes(searchQuery.toLowerCase()));
-      setBoardss(filteredBoards);
-    } else {
-      setBoardss(boards);
-    }
-  }, [searchQuery]);
-
-   const onRefresh = useCallback(async () => {
-    setRefreshing(true);
+  const getWorkspaces = async () => {
     try {
-      const response = await trelloAPI.getBoards('me');
-      setBoards(response);
-      setBoardss(response);
+      setRefreshing(true);
+      const response = await trelloAPI.getWorkspaces("me");
+      setWorkspaces(response);
+      setSelectedWorkspace(selectedWorkspace || response[0].id);
+
+      const responseBoards = await trelloAPI.getBoards("me");
+      setBoards(responseBoards);
     } catch (err: any) {
       console.log(err.message || 'Erreur inconnue');
     } finally {
       setRefreshing(false);
     }
+  }
+
+  const handleAdd = async (name: string) =>  {
+    if (visible === "Board") {
+      await trelloAPI.createBoard(name).then(() => {
+        Vibration.vibrate([0, 60, 0, 0]);
+      });
+      setVisible(null)
+      getWorkspaces();
+    } else if (visible === "Workspace") {
+      await trelloAPI.createWorkspace(name)
+      setVisible(null)
+      getWorkspaces();
+    }
+  }
+
+  const handleEdit = async (name: string) => {
+    if (edit) {
+      await trelloAPI.renameWorkspace(selectedWorkspace, name)
+      setEdit(null);
+      getWorkspaces();
+    }
+  }
+
+  const handleDeleteBoard = async (id: string) => {
+    await trelloAPI.deleteBoard(id).then(() => {
+      Vibration.vibrate([0, 60, 0, 0]);
+    });
+    getWorkspaces();
+  }
+
+  useEffect(() => {
+    if (searchQuery) {
+      const filteredBoards = boards.filter(board => board.name.toLowerCase().includes(searchQuery.toLowerCase()));
+      setBoards(filteredBoards);
+    } else {
+      getWorkspaces();
+    }
+  }, [searchQuery]);
+
+   useEffect(() => {
+    getWorkspaces();
    }, []);
 
   return (
       <PaperProvider>
-        <View style={styles.tooltip}>
-          <AddButton onPress={() => setVisible(true)} />
-        </View>
-
         <View>
           <Searchbar
             placeholder="Search..."
@@ -85,41 +95,56 @@ const Home = ({
           />
         </View>
 
+        {workspaces.length > 0 && (
+          <View style={styles.workspacesContainer}>
+            <Picker
+              selectedValue={selectedWorkspace || workspaces[0].id}
+              onValueChange={(itemValue) => setSelectedWorkspace(itemValue)}
+            >
+              {workspaces.map(workspace => (
+                <Picker.Item key={workspace.id} label={workspace.displayName} value={workspace.id} />
+              ))}
+            </Picker>
+          </View>
+        )}
+
         <ScrollView
           contentContainerStyle={{ flexGrow: 1 }}
           keyboardShouldPersistTaps="handled"
           refreshControl={
-            <RefreshControl refreshing={refreshing} onRefresh={onRefresh} colors={['#0079BF', '#D29034', '#519839', '#B04632', '#89609E']} />
+            <RefreshControl refreshing={refreshing} onRefresh={getWorkspaces} colors={['#0079BF', '#D29034', '#519839', '#B04632', '#89609E']} />
           }
         >
           <View style={styles.boardsContainer}>
             <Text style={styles.title}>Your board{boards.length > 1 ? 's' : ''}</Text>
-            {boardss.map(board => (
-              <Board
-                key={board.id}
-                board={board}
-                onPress={(id, name) => navigation.navigate('Board', { id, name })}
-                onLongPress={(id, name) => {
-                  Alert.alert(
-                    `Delete board "${name}"`,
-                    'Are you sure you want to delete this board?',
-                    [
-                      {
-                        text: 'Cancel',
-                        style: 'cancel',
-                      },
-                      {
-                        text: 'Delete',
-                        onPress: () => { handleDeleteBoard(id); },
-                      },
-                    ],
-                  );
-                }}
-              />
-            ))}
+            {boards.map(board => {
+              if (selectedWorkspace && board.idOrganization !== selectedWorkspace) return null;
+              return (
+                <Board
+                  key={board.id}
+                  board={board}
+                  onPress={(id, name) => navigation.navigate('Board', { id, name })}
+                  onLongPress={(id, name) => {
+                    Alert.alert(
+                      `Delete board "${name}"`,
+                      'Are you sure you want to delete this board?',
+                      [
+                        {
+                          text: 'Cancel',
+                          style: 'cancel',
+                        },
+                        {
+                          text: 'Delete',
+                          onPress: () => { handleDeleteBoard(id); },
+                        },
+                      ],
+                    );
+                  }}
+                />
+              )})}
 
-            {boardss.length === 0 &&
-              <Text style={{ fontSize: 20, color: 'gray', marginTop: 20 }}>
+            {((boards.length === 0 || boards.find(board => board.idOrganization === selectedWorkspace) === undefined) && !refreshing) &&
+              <Text style={{ fontSize: 20, color: 'gray', marginTop: 20, textAlign: 'center' }}>
               {searchQuery ? 'No boards found for this search :(' : 'You don\'t have any boards yet\nStart by creating one!'}
               </Text>
             }
@@ -127,10 +152,43 @@ const Home = ({
         </ScrollView>
 
         <Portal>
-          <Modal visible={visible} onDismiss={() => setVisible(false)} contentContainerStyle={containerStyle}>
-            <TextInput mode="outlined" placeholder="Board name..." autoFocus onSubmitEditing={(e) => handleAddBoard(e.nativeEvent.text)} />
+          <Modal visible={!!visible} onDismiss={() => setVisible(null)} contentContainerStyle={containerStyle}>
+            <TextInput mode="outlined" placeholder={`${visible} name...`} autoFocus onSubmitEditing={(e) => handleAdd(e.nativeEvent.text)} />
+          </Modal>
+
+          <Modal visible={!!edit} onDismiss={() => setEdit(null)} contentContainerStyle={containerStyle}>
+            <TextInput mode="outlined" autoFocus onSubmitEditing={(e) => handleEdit(e.nativeEvent.text)} value={edit} onChangeText={setEdit} />
           </Modal>
         </Portal>
+
+        <View style={styles.tooltip}>
+          <AddButton options={[
+            { icon: 'trash-can', label: 'Delete Workspace', onPress: () => {
+              Alert.alert(
+                'Delete Workspace',
+                'Are you sure you want to delete this workspace?',
+                [
+                  {
+                    text: 'Cancel',
+                    style: 'cancel',
+                  },
+                  {
+                    text: 'Delete',
+                    onPress: () => {
+                      trelloAPI.deleteWorkspace(selectedWorkspace).then(() => {
+                        Vibration.vibrate([0, 60, 0, 0]);
+                        getWorkspaces();
+                      })
+                    }
+                  },
+                ],
+              );
+            } },
+            { icon: 'pen', label: 'Edit Workspace', onPress: () => setEdit(workspaces.find(workspace => workspace.id === selectedWorkspace)?.displayName) },
+            { icon: "table", label: 'Create workspace', onPress: () => setVisible("Workspace") },
+            { icon: "bulletin-board", label: 'Create board', onPress: () => setVisible("Board") },
+          ]} />
+        </View>
       </PaperProvider>
   );
 };
@@ -161,6 +219,21 @@ const styles = StyleSheet.create({
     marginBottom: 10,
     width: '100%',
     marginLeft: 34,
+  },
+
+  workspacesContainer: {
+    marginHorizontal: 10,
+    marginTop: 20,
+    marginBottom: 10,
+    backgroundColor: '#fff',
+    borderRadius: 10,
+  },
+
+  fab: {
+    position: 'absolute',
+    top: WINDOW_HEIGHT - 80,
+    right: 15,
+    zIndex: 1,
   },
 });
